Put all the date-related items back in the Tag file. Date tests are working. Merging to work on int returns next.
